Closes #2606

+++ tests/RouterTest.php
@@ -5,6 +5,8 @@
  * @license https://github.com/slimphp/Slim/blob/4.x/LICENSE.md (MIT License)
  */
 
+declare(strict_types=1);
+
 namespace Slim\Tests;
 
 use FastRoute\RouteCollector;
@@ -19,15 +21,19 @@ use Slim\Tests\Mocks\InvocationStrategyTest;
 
 class RouterTest extends TestCase
 {
-    /** @var Router */
+    /**
+     * @var Router
+     */
     protected $router;
 
-    /** @var string */
+    /**
+     * @var null|string
+     */
     protected $cacheFile;
 
     public function tearDown()
     {
-        if (file_exists($this->cacheFile)) {
+        if ($this->cacheFile && file_exists($this->cacheFile)) {
             unlink($this->cacheFile);
         }
     }
@@ -365,7 +371,7 @@ class RouterTest extends TestCase
         };
         $route = $this->router->map($methods, $pattern, $callable)->setName('foo');
 
-        $cacheFile = dirname(__FILE__) . '/' . uniqid(microtime(true));
+        $cacheFile = dirname(__FILE__) . '/' . uniqid((string) microtime(true));
         $this->router->setCacheFile($cacheFile);
         $class = new ReflectionClass($this->router);
         $method = $class->getMethod('createDispatcher');
